Version 0.8
Greatly improved the speed when perl is available.  Now almost everything happens in
the perl interpreter.   Speed improvement example: on a file with 5408 tags:
645 second --> 72 second.  Limited the number of menu items to
configurable size both for time reasons and because gtk doesn't like menus which are
longer than the screen.
Added the TagsBase.pl and VIM.pm file to the distribution, those are not needed for
running the scripts but are usefull for debuging the perl parts.